ParserTesting2 extracted from ParserTesting
- ParserTesting only contains helpers requiring Parser and ParserContext.
